refactor: move config, loader, runner, validator into runtime/ package
