+++ builders/server/tests/runtime/test_runner.py
@@ -5,7 +5,7 @@ from typing import Any
 import pandas as pd
 import pytest
 
-import runner
+from runtime import runner
 
 # Top-level functions so they are picklable for multiprocessing
